Fix wrong merge
