+++ functions/index.js
@@ -7,7 +7,6 @@ const { initializeApp } = require('firebase-admin/app');
 initializeApp();
 
 const android = require('./android');
-const ios = require('./ios');
 const legacy = require('./legacy');
 
 const region = (functions.config().app && functions.config().app.region) || 'us-central1';
@@ -21,10 +20,6 @@ const { handleRequest, handleCheckRateLimits } = require('./handlers');
 
 exports.androidV1 = regionalFunctions.https.onRequest(async (req, res) =>
   handleRequest(req, res, android.createPayload),
-);
-
-exports.iOSV1 = regionalFunctions.https.onRequest(async (req, res) =>
-  handleRequest(req, res, ios.createPayload),
 );
 
 exports.sendPushNotification = regionalFunctions.https.onRequest(async (req, res) =>
