+++ functions/webapp.js
@@ -11,7 +11,6 @@ const fastify = require('fastify')({ logger: loggerConfig, trustProxy: true });
 const { handleRequest, handleCheckRateLimits } = require('./handlers');
 
 const android = require('./android');
-const ios = require('./ios');
 const legacy = require('./legacy');
 
 // Cloud Functions adapter
@@ -38,11 +37,6 @@ async function handleAndroidV1(request, reply) {
   return await handleRequest(req, res, android.createPayload);
 }
 
-async function handleIOSV1(request, reply) {
-  const { req, res } = createCloudFunctionsAdapter(request, reply);
-  return handleRequest(req, res, ios.createPayload);
-}
-
 async function handleSendPushNotification(request, reply) {
   const { req, res } = createCloudFunctionsAdapter(request, reply);
   return handleRequest(req, res, legacy.createPayload);
@@ -55,7 +49,6 @@ async function checkRateLimits(request, reply) {
 
 // Register routes
 fastify.post('/api/sendPush/android/v1', handleAndroidV1);
-fastify.post('/api/sendPush/iOS/v1', handleIOSV1);
 fastify.post('/api/sendPushNotification', handleSendPushNotification);
 fastify.post('/api/checkRateLimits', checkRateLimits);
